+++ src/app/api/orders/[orderId]/route.ts
@@ -1,11 +1,12 @@
 import { ERROR_CODE } from '@/lib/errors/errorCodes';
 import { requireActiveUser } from '@/app/api/_lib/auth';
 import { isApiMockEnabled } from '@/app/api/_lib/mock';
+import { expireUserOrders } from '@/app/api/_lib/order-expiration';
 import { fail, routeError, success } from '@/app/api/_lib/response';
 import { mockOrderDetailsMap } from '@/mocks/orders';
 
 import { orderIdSchema } from '../_lib/schemas';
-import { expireUserOrders, getOrder } from '../_lib/service';
+import { getOrder } from '../_lib/service';
 
 export async function GET(
   _request: Request,
