+++ src/app/api/payments/prepare/route.ts
@@ -1,9 +1,9 @@
 import type { NextRequest } from 'next/server';
 
 import { requireActiveUser } from '@/app/api/_lib/auth';
+import { expireUserOrders } from '@/app/api/_lib/order-expiration';
 import { routeError, success } from '@/app/api/_lib/response';
 import { validateBody } from '@/app/api/_lib/validation';
-import { expireUserOrders } from '@/app/api/orders/_lib/service';
 
 import { preparePaymentSchema } from '../_lib/schemas';
 import { preparePayment } from '../_lib/service';
